refactor global firestore data files (#717)
