+++ src/global-firestore-data/__test__/prepopulate-semesters.test.ts
@@ -1,4 +1,4 @@
-import { semesterEquals, getActiveSemesters } from '../semesters';
+import { semesterEquals, getActiveSemesters } from '../user-semesters';
 
 const semestersEqual = (
   expected: readonly FirestoreSemester[],
